Extract tables to components to avoid memoization
